requirements.txt
streamlit
pandas
joblib
scikit-learn
plotly
pillow
pytesseract